+++ tests/test_task.cc
@@ -34,6 +34,7 @@ void ExpectEq(const std::vector<std::vector<TaskOutput>>& actual,
       EXPECT_EQ(actual[i][j].task_input, expected[i][j].task_input);
       EXPECT_EQ(actual[i][j].image_width, expected[i][j].image_width);
       EXPECT_EQ(actual[i][j].image_height, expected[i][j].image_height);
+      EXPECT_EQ(actual[i][j].bit_depth, expected[i][j].bit_depth);
       EXPECT_EQ(actual[i][j].num_frames, expected[i][j].num_frames);
       EXPECT_EQ(actual[i][j].encoded_size, expected[i][j].encoded_size);
       EXPECT_EQ(actual[i][j].encoding_duration,
@@ -55,7 +56,7 @@ constexpr Subsampling kDef = Subsampling::kDefault;
 
 TEST(SplitByCodecSettingsAndAggregateByImageTest, Simple) {
   const std::vector<TaskOutput> results = {
-      {{{kWebp, kDef, /*effort=*/0, /*quality=*/0}, "img"}, 1, 2, 3, 0}};
+      {{{kWebp, kDef, /*effort=*/0, /*quality=*/0}, "img"}, 1, 2, 8, 3, 0}};
   const auto aggregate = SplitByCodecSettingsAndAggregateByImageAndQuality(
       results, /*quiet=*/false);
   ASSERT_EQ(aggregate.status, Status::kOk);
@@ -64,15 +65,16 @@ TEST(SplitByCodecSettingsAndAggregateByImageTest, Simple) {
 
 TEST(SplitByCodecSettingsAndAggregateByImageTest, Multiple) {
   const std::vector<TaskInput> single_inputs = {
-      {{kWebp, kDef, /*effort=*/0, /*quality=*/0}, "imgA"},
-      {{kWebp, kDef, /*effort=*/0, /*quality=*/0}, "imgB"},
-      {{kWebp, kDef, /*effort=*/1, /*quality=*/0}, "imgA"},
-      {{kWebp, kDef, /*effort=*/0, /*quality=*/100}, "imgA"},
-      {{kWebp2, kDef, /*effort=*/0, /*quality=*/0}, "imgA"}};
+      {{kWebp, kDef, /*effort=*/0, /*quality=*/0}, "A"},
+      {{kWebp, kDef, /*effort=*/0, /*quality=*/0}, "B"},
+      {{kWebp, kDef, /*effort=*/1, /*quality=*/0}, "A"},
+      {{kWebp, kDef, /*effort=*/0, /*quality=*/100}, "A"},
+      {{kWebp2, kDef, /*effort=*/0, /*quality=*/0}, "A"}};
   std::vector<TaskOutput> results;
   results.reserve(single_inputs.size() * 2);
   constexpr uint32_t kImageWidth = 8;
   constexpr uint32_t kImageHeight = 9;
+  constexpr uint32_t kImageDepth = 8;
   constexpr uint32_t kNumFrames = 1;
   size_t encoded_size = 1;
   double encoding_duration = 1;
@@ -85,6 +87,7 @@ TEST(SplitByCodecSettingsAndAggregateByImageTest, Multiple) {
     results.push_back({input,
                        kImageWidth,
                        kImageHeight,
+                       kImageDepth,
                        kNumFrames,
                        encoded_size,
                        encoding_duration++,
@@ -94,6 +97,7 @@ TEST(SplitByCodecSettingsAndAggregateByImageTest, Multiple) {
     results.push_back({input,
                        kImageWidth,
                        kImageHeight,
+                       kImageDepth,
                        kNumFrames,
                        encoded_size++,
                        encoding_duration++,
@@ -110,11 +114,11 @@ TEST(SplitByCodecSettingsAndAggregateByImageTest, Multiple) {
   ASSERT_EQ(aggregate.status, Status::kOk);
   ExpectEq(
       aggregate.value,
-      {{{{{kWebp, kDef, 0, 0}, "imgA"}, 8, 9, 1, 1u, 1.5, 1.5, 0.5, {20.0}},
-        {{{kWebp, kDef, 0, 100}, "imgA"}, 8, 9, 1, 4u, 7.5, 7.5, 3.5, {23.0}},
-        {{{kWebp, kDef, 0, 0}, "imgB"}, 8, 9, 1, 2u, 3.5, 3.5, 1.5, {21.0}}},
-       {{{{kWebp, kDef, 1, 0}, "imgA"}, 8, 9, 1, 3u, 5.5, 5.5, 2.5, {22.0}}},
-       {{{{kWebp2, kDef, 0, 0}, "imgA"}, 8, 9, 1, 5u, 9.5, 9.5, 4.5, {24.0}}}});
+      {{{{{kWebp, kDef, 0, 0}, "A"}, 8, 9, 8, 1, 1u, 1.5, 1.5, 0.5, {20.0}},
+        {{{kWebp, kDef, 0, 100}, "A"}, 8, 9, 8, 1, 4u, 7.5, 7.5, 3.5, {23.0}},
+        {{{kWebp, kDef, 0, 0}, "B"}, 8, 9, 8, 1, 2u, 3.5, 3.5, 1.5, {21.0}}},
+       {{{{kWebp, kDef, 1, 0}, "A"}, 8, 9, 8, 1, 3u, 5.5, 5.5, 2.5, {22.0}}},
+       {{{{kWebp2, kDef, 0, 0}, "A"}, 8, 9, 8, 1, 5u, 9.5, 9.5, 4.5, {24.0}}}});
 }
 
 }  // namespace